# app/public/admin-openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WOD Operations API",
        "description": "WOD Operations Ecommerce API",
        "contact": {
            "name": "John Doe",
            "email": "support@wayof.dev"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get list of categories.",
                "description": "Get list of categories.",
                "operationId": "f3673971d0f4ddb16b98f8d0b9a53a4d",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Store new product",
                "description": "This endpoint stores new product into database storage.",
                "operationId": "b9215817b901d3496022eef065a49b7e",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ViewCategoryDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/admin/categories/{category_id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get single category.",
                "description": "This endpoint returns a single category.",
                "operationId": "f4e0721222c05878c6120897d60f9518",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ViewCategoryDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update category contents.",
                "description": "This endpoint updates fields of category.",
                "operationId": "7caf79913d8508be8fba19d9daadc57b",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ViewCategoryDto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "FootprintDto": {
                "title": "Footprint",
                "description": "Footprint data.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "party": {
                        "type": "string",
                        "format": "string",
                        "example": "user"
                    },
                    "realm": {
                        "type": "string",
                        "format": "string",
                        "example": "front-office"
                    }
                },
                "type": "object"
            },
            "SignatureDto": {
                "title": "Signature",
                "description": "Signature data.",
                "properties": {
                    "at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "by": {
                        "schema": "Application\\Auth\\Dto\\FootprintDto",
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FootprintDto"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "StoreCategoryDto": {
                "title": "Category",
                "description": "Category data for storing new category.",
                "properties": {
                    "name": {
                        "description": "Name of the product.",
                        "type": "string",
                        "example": "Awesome Jacket"
                    },
                    "gender": {
                        "description": "Gender of the category.",
                        "type": "string",
                        "example": "male"
                    }
                },
                "type": "object"
            },
            "UpdateCategoryDto": {
                "title": "Category",
                "description": "Category data for updating category.",
                "properties": {
                    "name": {
                        "description": "Name of the product.",
                        "type": "string",
                        "example": "Awesome Jacket",
                        "nullable": true
                    },
                    "gender": {
                        "description": "Gender of the category.",
                        "type": "string",
                        "example": "male"
                    }
                },
                "type": "object"
            },
            "ViewCategoryDto": {
                "title": "Category",
                "description": "Category response data.",
                "properties": {
                    "incrementalId": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Men Jackets"
                    },
                    "gender": {
                        "description": "Gender of the category.",
                        "type": "string",
                        "example": "male"
                    },
                    "created": {
                        "$ref": "#/components/schemas/SignatureDto"
                    },
                    "updated": {
                        "$ref": "#/components/schemas/SignatureDto"
                    },
                    "deleted": {
                        "schema": "Application\\Auth\\Dto\\SignatureDto",
                        "nullable": true,
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SignatureDto"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    }
}